# auth-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Este é o serviço de autenticação do sistema de microserviços.",
        "title": "Auth Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/usuarios": {
            "post": {
                "description": "Cadastra um novo usuário no sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuários"
                ],
                "summary": "Cadastrar um novo usuário",
                "parameters": [
                    {
                        "description": "Dados do usuário",
                        "name": "usuario",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Usuario"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Usuário cadastrado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/schema.UsuarioCreatedEnvelope"
                        }
                    },
                    "400": {
                        "description": "Requisição inválida",
                        "schema": {
                            "$ref": "#/definitions/error.AppError"
                        }
                    },
                    "409": {
                        "description": "Conflito: usuário já existe",
                        "schema": {
                            "$ref": "#/definitions/error.AppError"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/error.AppError"
                        }
                    }
                }
            }
        },
        "/usuarios/logado": {
            "get": {
                "description": "Busca as informações do usuário logado a partir do token JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuários"
                ],
                "summary": "Buscar informações do usuário logado",
                "responses": {
                    "200": {
                        "description": "Usuário encontrado",
                        "schema": {
                            "$ref": "#/definitions/schema.UsuarioEnvelope"
                        }
                    },
                    "401": {
                        "description": "Token inválido ou expirado",
                        "schema": {
                            "$ref": "#/definitions/error.AppError"
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado",
                        "schema": {
                            "$ref": "#/definitions/error.AppError"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/error.AppError"
                        }
                    }
                }
            }
        },
        "/usuarios/login": {
            "post": {
                "description": "Realiza o login de um usuário e retorna um token JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuários"
                ],
                "summary": "Realizar login de um usuário",
                "parameters": [
                    {
                        "description": "Dados de login do usuário",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/schema.LoginResponseEnvelope"
                        }
                    },
                    "400": {
                        "description": "Requisição inválida",
                        "schema": {
                            "$ref": "#/definitions/error.AppError"
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas",
                        "schema": {
                            "$ref": "#/definitions/error.AppError"
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado",
                        "schema": {
                            "$ref": "#/definitions/error.AppError"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/error.AppError"
                        }
                    }
                }
            }
        },
        "/usuarios/{id}": {
            "put": {
                "description": "Altera as informações do usuário logado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuários"
                ],
                "summary": "Alterar informações do usuário",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do usuário",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do usuário",
                        "name": "usuario",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Usuario"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário atualizado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/schema.UsuarioEnvelope"
                        }
                    },
                    "400": {
                        "description": "Requisição inválida",
                        "schema": {
                            "$ref": "#/definitions/error.AppError"
                        }
                    },
                    "401": {
                        "description": "Token inválido ou expirado",
                        "schema": {
                            "$ref": "#/definitions/error.AppError"
                        }
                    },
                    "403": {
                        "description": "Acesso negado",
                        "schema": {
                            "$ref": "#/definitions/error.AppError"
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado",
                        "schema": {
                            "$ref": "#/definitions/error.AppError"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/error.AppError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Login": {
            "type": "object",
            "required": [
                "email",
                "senha"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "senha": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
        "domain.Usuario": {
            "type": "object",
            "required": [
                "email",
                "nome",
                "senha"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "senha": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 6
                }
            }
        },
        "error.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "ex.: USER_NOT_FOUND",
                    "type": "string"
                },
                "message": {
                    "description": "msg “safe” para cliente",
                    "type": "string"
                }
            }
        },
        "schema.LoginResponseEnvelope": {
            "description": "Resposta do login do usuário",
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "token": {
                            "type": "string"
                        },
                        "usuario": {
                            "$ref": "#/definitions/domain.Usuario"
                        }
                    }
                }
            }
        },
        "schema.UsuarioCreatedEnvelope": {
            "description": "Resposta do usuário criado",
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "schema.UsuarioEnvelope": {
            "description": "Resposta do usuário criado",
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "usuario": {
                            "$ref": "#/definitions/domain.Usuario"
                        }
                    }
                }
            }
        }
    }
}